refactor : domain 모듈 분리
- domain model, repository 추가
