--- app/src/main/java/com/kky/pokedex/local/entity/Like.kt
@@ -1,10 +1,0 @@
-package com.kky.pokedex.local.entity
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "like")
-data class Like(
-    @PrimaryKey(autoGenerate = false)
-    val id: Int
-)
