+++ app/src/main/java/com/kky/pokedex/ui/MainViewModel.kt
@@ -4,7 +4,7 @@ import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.kky.pokedex.domain.model.Pokemon
-import com.kky.pokedex.repository.PokemonRepository
+import com.kky.pokedex.domain.repository.PokemonRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
