+++ app/src/main/java/com/kky/pokedex/ui/detail/PokemonDetailViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.kky.pokedex.domain.model.PokemonDetail
-import com.kky.pokedex.repository.PokemonRepository
+import com.kky.pokedex.domain.repository.PokemonRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
