+++ build.gradle.kts
@@ -4,4 +4,5 @@ plugins {
     alias(libs.plugins.jetbrainsKotlinAndroid) apply false
     alias(libs.plugins.hiltPlugin) apply false
     alias(libs.plugins.ksp) apply false
+    alias(libs.plugins.jetbrainsKotlinJvm) apply false
 }
