+++ domain/src/main/java/com/kky/pokedex/domain/MyClass.kt
@@ -1,0 +1,4 @@
+package com.kky.pokedex.domain
+
+class MyClass {
+}
