+++ domain/src/main/java/com/kky/pokedex/domain/model/Pokemon.kt
